Refactored: Datadog::Encoding --> Datadog::Core::Encoding
