+++ spec/ddtrace/transport/http/api_spec.rb
@@ -19,14 +19,14 @@ RSpec.describe Datadog::Transport::HTTP::API do
         expect(v4).to be_a_kind_of(Datadog::Transport::HTTP::Traces::API::Spec)
         expect(v4.traces).to be_a_kind_of(Datadog::Transport::HTTP::Traces::API::Endpoint)
         expect(v4.traces.service_rates?).to be true
-        expect(v4.traces.encoder).to be Datadog::Encoding::MsgpackEncoder
+        expect(v4.traces.encoder).to be Datadog::Core::Encoding::MsgpackEncoder
       end
 
       defaults[described_class::V3].tap do |v3|
         expect(v3).to be_a_kind_of(Datadog::Transport::HTTP::Traces::API::Spec)
         expect(v3.traces).to be_a_kind_of(Datadog::Transport::HTTP::Traces::API::Endpoint)
         expect(v3.traces.service_rates?).to be false
-        expect(v3.traces.encoder).to be Datadog::Encoding::MsgpackEncoder
+        expect(v3.traces.encoder).to be Datadog::Core::Encoding::MsgpackEncoder
       end
     end
