+++ spec/ddtrace/transport/io/client_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Datadog::Transport::IO::Client do
   subject(:client) { described_class.new(out, encoder) }
 
   let(:out) { instance_double(IO) }
-  let(:encoder) { instance_double(Datadog::Encoding::Encoder) }
+  let(:encoder) { instance_double(Datadog::Core::Encoding::Encoder) }
 
   describe '#initialize' do
     it { is_expected.to be_a_kind_of Datadog::Transport::Statistics }
